Support for Optional SensorMeta
Updated getFromDict() to return None when the key path does not resolve
recursively. Previously this would result in "TypeError: 'NoneType' object is
not subscriptable". This change is needed to support SenorMeta that does not
exist on all car types. Specifically this change is needed for a 911 that does
not support "remainingRanges.electricalRange.distance".