Propose to remove the word 'sensor' from sensor names
This is a proposal for a minor cleanup. Currently, sensors have the name "foo sensor". For example, the range sensor gets the entity id sensor.model_range_sensor. I propose to remove the word sensor to shorten the entity id of each sensor, as it is redundant information.

This change does not affect those who have already configured this component in hass. The sensor names will remain the same. It only affects new installations.